requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
streamlit==1.32.2
selenium>=4.20.0
undetected-chromedriver>=3.5.5
langchain==0.1.13
langchain-community==0.0.29
Pillow==10.2.0
python-dotenv==1.0.1
pydantic==2.6.4
requests==2.31.0
cffi==1.17.1
cryptography==44.0.0
passlib==1.7.4
bcrypt==4.1.2
pygments==2.17.2
pyautogui==0.9.54
fpdf2==2.7.9
beautifulsoup4==4.12.3
